Drive engine.test.ts's strike/carryover/rangestrike wrappers through a real TitanGame
The wrappers previously called gameBattle's mutators via
`.call({ activeBattle: battle } as unknown as TitanGame, ...)`, a synthetic
`this` that would silently no-op instead of type-erroring if those mutators
ever grew to touch more of TitanGame than activeBattle. game.test.ts already
had a createActionContext helper that maps commit("fooBar", ...) to a real
TitanGame's mFooBar method the same way the Vuex module does; extracting it
(plus createGetters and newGame) into game.testUtils.ts lets engine.test.ts
reuse it against a real TitanGame instance instead of a cast object.
